+++ src/Providers/CommandServiceProvider.php
@@ -2,6 +2,7 @@
 
 namespace Tapestry\Providers;
 
+use Tapestry\Console\Commands\MakeCommand;
 use Tapestry\Tapestry;
 use Tapestry\Console\Application;
 use Tapestry\Console\Commands\InitCommand;
@@ -48,6 +49,12 @@ class CommandServiceProvider extends AbstractServiceProvider
                 \Symfony\Component\Finder\Finder::class,
             ]);
 
+        $container->add(MakeCommand::class)
+            ->withArguments([
+                \Symfony\Component\Filesystem\Filesystem::class,
+                \Symfony\Component\Finder\Finder::class,
+            ]);
+
         $container->share(Application::class)
             ->withArguments([
                 Tapestry::class,
@@ -55,6 +62,7 @@ class CommandServiceProvider extends AbstractServiceProvider
                     $container->get(InitCommand::class),
                     $container->get(BuildCommand::class),
                     $container->get(SelfUpdateCommand::class),
+                    $container->get(MakeCommand::class)
                 ],
             ]);
     }