+++ tests/ViewFileTraitTest.php
@@ -39,7 +39,7 @@ class ViewFileTraitTest extends CommandTestBase
      */
     public function testExcerptHelper()
     {
-        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/mocks/TestExcerptFile.md');
+        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/Mocks/TestExcerptFile.md');
 
         $this->assertEquals('Lorem ipsum dolor sit amet, consectetur&hellip;', $viewFile->getExcerpt());
         $this->assertEquals('Lorem ipsum dolor sit amet, consectetur', $viewFile->getExcerpt(50, ''));
@@ -49,13 +49,13 @@ class ViewFileTraitTest extends CommandTestBase
         $this->assertEquals('Lorem', $viewFile->getExcerpt(10, ''));
         $this->assertEquals('Lorem', $viewFile->getExcerpt(10, null));
 
-        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/mocks/TestFile.md');
+        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/Mocks/TestFile.md');
         $this->assertEquals('This is a test file...', $viewFile->getExcerpt(50, null));
     }
 
     public function testGetContentHelper()
     {
-        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/mocks/TestFile.md');
+        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/Mocks/TestFile.md');
         $this->assertEquals('This is a test file...', trim($viewFile->getContent()));
 
         $viewFile->getFile()->setData(['content' => 'TEST test TEST 1234']);
@@ -64,7 +64,7 @@ class ViewFileTraitTest extends CommandTestBase
 
     public function testGetPermalinkHelper()
     {
-        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/mocks/TestExcerptFile.md');
+        $viewFile = $this->mockViewFile($this->mockTapestry(),__DIR__ . '/Mocks/TestExcerptFile.md');
         $this->assertEquals('/testexcerptfile/index.md', $viewFile->getPermalink()); // this has the extension md becuase its not passed through any renderers
 
         $viewFile->getFile()->setData(['permalink' => '/folder1/folder2/folder3/test.html']);
@@ -75,7 +75,7 @@ class ViewFileTraitTest extends CommandTestBase
     {
         $viewFile = $this->mockViewFile(
             $this->mockTapestry(null, ['site' => ['url' => 'http://www.example.com']]),
-            __DIR__ . '/mocks/TestExcerptFile.md'
+            __DIR__ . '/Mocks/TestExcerptFile.md'
         );
 
         $this->assertEquals('http://www.example.com/testexcerptfile/', $viewFile->getUrl());
@@ -88,7 +88,7 @@ class ViewFileTraitTest extends CommandTestBase
     {
         $viewFile = $this->mockViewFile(
             $this->mockTapestry(null, ['site' => ['url' => 'http://www.example.com']]),
-            __DIR__ . '/mocks/TestExcerptFile.md'
+            __DIR__ . '/Mocks/TestExcerptFile.md'
         );
 
         $this->assertEquals(false, $viewFile->isPaginated());
@@ -104,7 +104,7 @@ class ViewFileTraitTest extends CommandTestBase
     {
         $viewFile = $this->mockViewFile(
             $this->mockTapestry(null, ['site' => ['url' => 'http://www.example.com']]),
-            __DIR__ . '/mocks/TestExcerptFile.md'
+            __DIR__ . '/Mocks/TestExcerptFile.md'
         );
 
         $this->assertEquals(false, $viewFile->hasPreviousNext());
@@ -120,7 +120,7 @@ class ViewFileTraitTest extends CommandTestBase
     {
         $viewFile = $this->mockViewFile(
             $this->mockTapestry(),
-            __DIR__ . '/mocks/TestFile.md'
+            __DIR__ . '/Mocks/TestFile.md'
         );
 
         $this->assertInstanceOf(\DateTime::class, $viewFile->getDate());